Fix issue with spaces and less than operator
The solution to the spaces problem is to wrap the argument in quotation marks (this only works if there is only one pair of quotation marks!)
The problem with the less than character is that CALL does not allow the less than character (see https://docs.microsoft.com/en-us/windows-server/administration/windows-commands/call):  "Do not use pipes (|) or redirection symbols (< or >) with call." 
The solution could be to replace that character with a placeholder, and to replace that placeholder with a less than character in seriatim.py. I propose using LESSTHAN rather than the more common "LT" because "LT" is common enough in natural language to create problems if it is replaced with "<".